refactor: 배치 작업 초기화 및 게시글 참조 ID 누락 필드 추가
